:sparkles: FEAT: Sound added for click and hover events
