--- src/components/AuthorInfo/Achievements.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const Achievements = () => {
-  return <div>
-
-
-
-
-
-
-
-      
-  </div>;
-};
-
-export default Achievements;
