--- src/components/AuthorInfo/index.js
@@ -1,2 +1,0 @@
-import AuthorInfo from './AuthorInfo'
-export default AuthorInfo
