+++ src/components/BlogPreview/BlogPreview.style.js
@@ -3,26 +3,37 @@ import { Flex } from '@portfolio-ui/';
 import { Link } from 'gatsby';
 
 export const BlogItem = styled(Link)`
-  color: ${props => props.theme.primaryBlack};
   display: block;
   margin-bottom: 70px;
-  box-shadow:  ${props => props.theme.cardShadow};
+  box-shadow: ${props => props.theme.cardShadow};
   padding: 2rem 3rem;
   border-radius: 10px;
   border: 1px solid
     ${props =>
       props.theme.dark ? props.theme.primaryColor : 'hsla(0, 0%, 50%, 0.1)'};
-  background: ${props => props.theme.secondaryColor};
+  background: ${props => props.theme.cardBg};
   transition: 0.3s;
 
   &:hover {
     transform: translate(0, -10px);
+    h1 {
+      color: ${p => p.theme.primaryColor};
+    }
+  }
+
+  @media ${props => props.theme.media.tablet} {
+    padding: 1.3rem 1rem;
+    margin-bottom : 50px;
   }
 `;
 
 export const DateTime = styled.p`
-  font-family: 'Ubuntu', sans-serif;
+  font-family: ${p => p.theme.UbuntuFontFamily};
   padding: 1rem 0;
+
+  @media ${props => props.theme.media.tablet} {
+    font-size: 1.1rem;
+  }
 `;
 
 export const BlogTitle = styled.h1`
@@ -33,9 +44,9 @@ export const BlogTitle = styled.h1`
 `;
 
 export const Description = styled.p`
-  padding: .5rem 0;
-  font-size : 1.1rem;
-  font-weight : 400;
+  padding: 0.5rem 0;
+  font-size: 1.1rem;
+  font-weight: 400;
 `;
 
 export const Tags = styled(Flex)`
@@ -44,6 +55,7 @@ export const Tags = styled(Flex)`
     display: block;
     font-weight: 500;
     background: ${props => props.theme.tagColor};
+    color: ${p => p.theme.primaryText};
     padding: 0.4rem 0.6rem;
     font-size: 0.75rem;
     border-radius: 3rem;
