+++ src/components/Footer/Footer.js
@@ -5,7 +5,7 @@ import { FooterWrapper } from './Footer.style';
 export default function Footer() {
   return (
     <FooterWrapper>
-     © Aadarsha {new Date().getFullYear()}, Built with
+      © Aadarsha {new Date().getFullYear()}, Built with
       {` `}
       <ExternalLink href="https://www.gatsbyjs.org">Gatsby</ExternalLink>{' '}
     </FooterWrapper>
