+++ src/components/Home/index.js
@@ -1,0 +1,2 @@
+import AuthorInfo from './AuthorInfo';
+export default AuthorInfo;
