+++ src/components/Navbar/desktop/NavDesktop.js
@@ -2,11 +2,13 @@ import React from 'react';
 
 import { NavItem, NavItems, Logo } from './NavDesktop.style';
 import NavDesktopLinks from './NavDesktopLinks';
+import { useSfx } from '@hooks/use-sfx';
 
 const NavDesktop = () => {
+  const { playClick } = useSfx();
   return (
     <>
-      <Logo to="/">
+      <Logo to="/" onClick={playClick}>
         <span>aada</span>
         rsha
       </Logo>
