+++ src/components/PageNotFound/PageNotFound.style.js
@@ -4,7 +4,7 @@ export const StyledPageNotFound = styled.div`
   height: 65vh;
 
   h1 {
-    font-family: Montserrat, sans-serif;
+    font-family: ${p => p.theme.UbuntuFontFamily}, sans-serif;
     font-weight: 900 !important;
     font-size: 35px;
   }
