+++ src/pages/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Layout from '@components/Layout';
 import SEO from '@components/Seo';
-import AuthorInfo from '@components/AuthorInfo';
+import AuthorInfo from '@src/components/Home';
 
 export default function IndexPage() {
   return (
