+++ src/pages/projects.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import { Link } from 'gatsby';
-
 import Layout from '@components/Layout';
 import SEO from '@components/Seo';
 import ProjectPreview from '@components/ProjectPreview';
